# docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "This is a retail management system server.",
        "title": "Retail Pro API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8000",
    "basePath": "/api/v1",
    "paths": {
...
        "/products": {
            "get": {
                "description": "Retrieves a list of all products available in the system. (No user filtering currently)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "products"
                ],
                "summary": "Get all products",
                "responses": {
                    "200": {
                        "description": "Successfully retrieved products",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/routes_product.Product"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_yantology_simple-pos_pkg_dto.MessageResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Creates a new product associated with the authenticated user.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "products"
                ],
                "summary": "Create a new product",
                "parameters": [
                    {
                        "description": "Product details",
                        "name": "product",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/routes_product.CreateProduct"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Product created successfully",
                        "schema": {
                            "$ref": "#/definitions/routes_product.Product"
                        }
                    },
                    "400": {
                        "description": "Invalid request data",
                        "schema": {
                            "$ref": "#/definitions/github_com_yantology_simple-pos_pkg_dto.MessageResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized: User ID not found in context",
                        "schema": {
                            "$ref": "#/definitions/github_com_yantology_simple-pos_pkg_dto.MessageResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_yantology_simple-pos_pkg_dto.MessageResponse"
                        }
                    }
                }
            }
        },
        "/products/category/{categoryID}": {
            "get": {
                "description": "Retrieves all products belonging to a specific Category ID.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "products"
                ],
                "summary": "Get products by Category ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Category ID",
                        "name": "categoryID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully retrieved products",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/routes_product.Product"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid Category ID format",
                        "schema": {
                            "$ref": "#/definitions/github_com_yantology_simple-pos_pkg_dto.MessageResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_yantology_simple-pos_pkg_dto.MessageResponse"
                        }
                    }
                }
            }
        },
        "/products/{id}": {
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Updates an existing product by its ID. User must own the product.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "products"
                ],
                "summary": "Update an existing product",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Product ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Updated product details",
                        "name": "product",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/routes_product.UpdateProduct"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Product updated successfully",
                        "schema": {
                            "$ref": "#/definitions/routes_product.Product"
                        }
                    },
                    "400": {
                        "description": "Invalid request data or ID format",
                        "schema": {
                            "$ref": "#/definitions/github_com_yantology_simple-pos_pkg_dto.MessageResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized: User ID not found in context or not owner",
                        "schema": {
                            "$ref": "#/definitions/github_com_yantology_simple-pos_pkg_dto.MessageResponse"
                        }
                    },
                    "404": {
                        "description": "Product not found",
                        "schema": {
                            "$ref": "#/definitions/github_com_yantology_simple-pos_pkg_dto.MessageResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_yantology_simple-pos_pkg_dto.MessageResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Deletes a product by its ID. User must own the product.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "products"
                ],
                "summary": "Delete a product",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Product ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Product deleted successfully",
                        "schema": {
                            "$ref": "#/definitions/github_com_yantology_simple-pos_pkg_dto.MessageResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid product ID format",
                        "schema": {
                            "$ref": "#/definitions/github_com_yantology_simple-pos_pkg_dto.MessageResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized: User ID not found in context or not owner",
                        "schema": {
                            "$ref": "#/definitions/github_com_yantology_simple-pos_pkg_dto.MessageResponse"
                        }
                    },
                    "404": {
                        "description": "Product not found",
                        "schema": {
                            "$ref": "#/definitions/github_com_yantology_simple-pos_pkg_dto.MessageResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_yantology_simple-pos_pkg_dto.MessageResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "github_com_yantology_simple-pos_pkg_dto.MessageResponse": {
            "description": "Generic message response model",
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "Operation completed successfully"
                }
            }
        },
...
        "routes_product.CreateProduct": {
            "description": "Create product request model",
            "type": "object",
            "required": [
                "category_id",
                "name",
                "price"
            ],
            "properties": {
                "category_id": {
                    "description": "Changed from string to int",
                    "type": "integer",
                    "example": 1
                },
                "is_available": {
                    "type": "boolean",
                    "example": true
                },
                "name": {
                    "type": "string",
                    "example": "Wireless Mouse"
                },
                "price": {
                    "type": "number",
                    "example": 250000
                }
            }
        },
        "routes_product.Product": {
            "description": "Product model",
            "type": "object",
            "properties": {
                "category_id": {
                    "description": "Changed from string to int",
                    "type": "integer",
                    "example": 1
                },
                "created_at": {
                    "type": "string",
                    "example": "2025-04-25T15:04:05Z07:00"
                },
                "id": {
                    "description": "Changed from string to int",
                    "type": "integer",
                    "example": 1
                },
                "is_available": {
                    "type": "boolean",
                    "example": true
                },
                "name": {
                    "type": "string",
                    "example": "Laptop Pro"
                },
                "price": {
                    "type": "number",
                    "example": 15000000
                },
                "updated_at": {
                    "type": "string",
                    "example": "2025-04-25T15:04:05Z07:00"
                },
                "user_id": {
                    "description": "Changed from string to int",
                    "type": "integer",
                    "example": 1
                }
            }
        },
        "routes_product.UpdateProduct": {
            "description": "Update product request model",
            "type": "object",
            "required": [
                "category_id",
                "name",
                "price"
            ],
            "properties": {
                "category_id": {
                    "description": "Changed from string to int",
                    "type": "integer",
                    "example": 2
                },
                "is_available": {
                    "type": "boolean",
                    "example": false
                },
                "name": {
                    "type": "string",
                    "example": "Laptop Pro X"
                },
                "price": {
                    "type": "number",
                    "example": 16500000
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}